Changes; - PacketLossAtIntervals script now produces results that allow congestion window to be monitored over time

+++ PacketCapture/Effects/PacketLossTimeDrop.py
@@ -1,10 +1,10 @@
 import time
 from Effects.Effect import Effect
-
+from scapy.all import *
 
 class PacketLossTimeDrop(Effect):
 
-    def __init__(self,
+    def __init__(self, drop_interval,
                  accept_packets=True,
                  show_output=True,
                  graphing=False,
@@ -21,12 +21,15 @@ class PacketLossTimeDrop(Effect):
         self.total_packets = 0
         self.time_since_last_drop = -1
 
-        self.DROP_INTERVAL_SECONDS = 1
+        self.DROP_INTERVAL_SECONDS = drop_interval
 
     def custom_effect(self, packet):
         """This function will issue packet loss,
            a percentage is defined and anything
            lower is dropped and anything higher is accepted"""
+
+        # This is the default port for Iperf
+        INCOMMING_PORT = 5001
 
         # Initial set
         if self.time_since_last_drop is -1:
@@ -36,11 +39,16 @@ class PacketLossTimeDrop(Effect):
             elapsed = time.time() - self.time_since_last_drop
 
             if elapsed > self.DROP_INTERVAL_SECONDS:
-                self.time_since_last_drop = time.time()
-                packet.drop()
-                print('DROP')
+
+                try:
+                    pkt = IP(packet.get_payload())
+                    if pkt.sport is not INCOMMING_PORT:
+                        self.time_since_last_drop = time.time()
+                        packet.drop()
+                except:
+                    pass
             else:
                 self.accept(packet)
 
     def print_stats(self):
-        print("[*] " + str(self.total_packets), end='\r')
+        self.print("[*] " + str(self.total_packets), end='\r')
